Add Devise and create a user
This commit is the second try of adding Devise to the application.  Now
looking back, the first time was perfectly fine, it was the docker image
that was missing up but now there is a way to solve that.  This commit
is just redoing what we undid when we thought Devise was the culprit.
This also added a site controller so we can see something other than the
Rails splash screen now.